Move node filtering functions to download.py
